+++ src/management/projects.spec.ts
@@ -1,3 +1,4 @@
+import * as clc from "colorette";
 import { expect } from "chai";
 import * as sinon from "sinon";
 import nock from "../test/helpers/nock";
@@ -438,7 +439,42 @@ describe("Project management", () => {
 
         expect(err).to.be.an.instanceOf(FirebaseError);
         expect(err?.message).to.equal(
-          `Failed to add Firebase to Google Cloud Platform project because your account has not accepted the Firebase Terms of Service. Please accept the Terms of Service in the Firebase console at ${api.consoleOrigin()} and try again.`,
+          `Failed to add Firebase to Google Cloud Platform project ${clc.bold(PROJECT_ID)} because your account has not accepted the Firebase Terms of Service. Please accept the Terms of Service in the Firebase console at ${api.consoleOrigin()} and try again.`,
+        );
+        expect(err?.original).to.be.an.instanceOf(FirebaseError);
+        expect(nock.isDone()).to.be.true;
+        expect(pollOperationStub).to.be.not.called;
+      });
+
+      it("should reject with the generic error if a 403 is not caused by unaccepted TOS", async () => {
+        nock(api.firebaseApiOrigin())
+          .post(`/v1beta1/projects/${PROJECT_ID}:addFirebase`)
+          .reply(403, {
+            error: {
+              code: 403,
+              message:
+                "Permission 'firebase.projects.update' denied on resource (or it may not exist).",
+              status: "PERMISSION_DENIED",
+              details: [
+                {
+                  "@type": "type.googleapis.com/google.rpc.ErrorInfo",
+                  reason: "IAM_PERMISSION_DENIED",
+                  domain: "firebase.googleapis.com",
+                },
+              ],
+            },
+          });
+
+        let err: FirebaseError | undefined;
+        try {
+          await projectManager.addFirebaseToCloudProject(PROJECT_ID);
+        } catch (e: unknown) {
+          err = e as FirebaseError;
+        }
+
+        expect(err).to.be.an.instanceOf(FirebaseError);
+        expect(err?.message).to.equal(
+          "Failed to add Firebase to Google Cloud Platform project. See firebase-debug.log for more info.",
         );
         expect(err?.original).to.be.an.instanceOf(FirebaseError);
         expect(nock.isDone()).to.be.true;